Rmarkdown template for an augmented version of the basic pandoc latex template

Rmarkdown template for an augmented version of the basic pandoc latex template

pdf_article(..., number_sections = TRUE,
  md_extensions = c("-autolink_bare_uris"), indent = TRUE,
  marginals = TRUE, author_block = TRUE, noindentafter = TRUE,
  subfigs = TRUE, template = NULL, csl = NULL, colorlinks = TRUE)

Arguments

...
arguments passed down to [bookdown::pdf_document2()]

number_sections
whether sections should be numbered

md_extensions
which markdown extensions to add

indent
whether to use paragraph indentation or not. In the latter case paragraphs are separated by vertical space as in the default rmarkdown template.

marginals
whether to augment latex template with marginals (header and footer) code

author_block
whether to augment latex template with author blocks. The latex package authblk is required.

noindentafter
whether to augment latex template with the noindentafter latex package to suppress indentation after certain environments. Always set to `FALSE` is `indent = FALSE`.

subfigs
whether to use the subfig package for subfig support

template
template for latex output. If `NULL`, the default latex template will be used (obtained from calling `pandoc -D latex`).

csl
bibliography style in the `.csl` format

colorlinks
should cross-references and links be colored?

Value

R Markdown output format to pass to [rmarkdown::render()]

Examples

library(rmarkdown)
library(bookdown)
rmarkdown::draft("my-article.Rmd",
                 template = "pdf_article",
                 package = "rmdplugr",
                 edit = FALSE)

#> Error in rmarkdown::draft("my-article.Rmd", template = "pdf_article",     package = "rmdplugr", edit = FALSE): The template 'pdf_article' was not found in the rmdplugr package